Splitting the utilities file into multiple - PART1
