+++ verticapy/io/__init__.py
@@ -14,3 +14,6 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
+from .utilities import *
+from .parsers.csv import read_csv
+from .parsers.json import read_json
