+++ verticapy/io/utilities.py
@@ -33,7 +33,7 @@ from verticapy.utils._decorators import (
 from verticapy.utils._toolbox import *
 from verticapy.jupyter._javascript import datatables_repr
 from verticapy.errors import *
-import verticapy.plotting._colors as vp_colors
+from .parsers.csv import read_csv
 
 # Other Modules
 import pandas as pd
@@ -47,91 +47,6 @@ except:
 #
 # Utilities Functions
 #
-
-
-@save_verticapy_logs
-def compute_flextable_keys(flex_name: str, usecols: list = []):
-    """
-Computes the flex table keys and returns the predicted data types.
-
-Parameters
-----------
-flex_name: str
-    Flex table name.
-usecols: list, optional
-    List of columns to consider.
-
-Returns
--------
-List of tuples
-    List of virtual column names and their respective data types.
-    """
-    executeSQL(
-        query=f"""
-            SELECT 
-                /*+LABEL('utilities.compute_flex_table_keys')*/
-                compute_flextable_keys('{flex_name}');""",
-        title="Guessing flex tables keys.",
-    )
-    usecols_str = [
-        "'" + str(column).lower().replace("'", "''") + "'" for column in usecols
-    ]
-    usecols_str = ", ".join(usecols_str)
-    where = f" WHERE LOWER(key_name) IN ({usecols_str})" if (usecols) else ""
-    result = executeSQL(
-        query=f"""
-            SELECT 
-                /*+LABEL('utilities.compute_flex_table_keys')*/
-                key_name,
-                data_type_guess 
-            FROM {flex_name}_keys{where}""",
-        title="Guessing the data types.",
-        method="fetchall",
-    )
-    return result
-
-
-@save_verticapy_logs
-def compute_vmap_keys(
-    expr: Union[str, str_sql], vmap_col: str, limit: int = 100,
-):
-    """
-Computes the most frequent keys in the input VMap.
-
-Parameters
-----------
-expr: str / vDataFrame
-    Input expression. You can also specify a vDataFrame or a customized 
-    relation, but you must enclose it with an alias. For example, "(SELECT 1) x" 
-    is allowed, whereas "(SELECT 1)" and "SELECT 1" are not.
-vmap_col: str
-    VMap column.
-limit: int, optional
-    Maximum number of keys to consider.
-
-Returns
--------
-List of tuples
-    List of virtual column names and their respective frequencies.
-    """
-    from verticapy import vDataFrame
-
-    vmap = quote_ident(vmap_col)
-    if isinstance(expr, vDataFrame):
-        assert expr[vmap_col].isvmap(), ParameterError(
-            f"Virtual column {vmap_col} is not a VMAP."
-        )
-        expr = expr.__genSQL__()
-    result = executeSQL(
-        (
-            "SELECT /*+LABEL('utilities.compute_vmap_keys')*/ keys, COUNT(*) FROM "
-            f"(SELECT MAPKEYS({vmap}) OVER (PARTITION BEST) FROM {expr})"
-            f" VERTICAPY_SUBTABLE GROUP BY 1 ORDER BY 2 DESC LIMIT {limit};"
-        ),
-        title="Getting vmap most occurent keys.",
-        method="fetchall",
-    )
-    return result
 
 
 def create_schema(
@@ -830,88 +745,6 @@ pandas_to_vertica : Ingests a pandas DataFrame into the Vertica database.
             return total_rows
 
 
-def isflextable(table_name: str, schema: str):
-    """
-Checks if the input relation is a flextable.
-
-Parameters
-----------
-table_name: str
-    Name of the table to check.
-schema: str
-    Table schema.
-
-Returns
--------
-bool
-    True if the relation is a flex table.
-    """
-    table_name = quote_ident(table_name)[1:-1]
-    schema = quote_ident(schema)[1:-1]
-    sql = (
-        f"SELECT is_flextable FROM v_catalog.tables WHERE table_name = '{table_name}' AND "
-        f"table_schema = '{schema}' AND is_flextable LIMIT 1;"
-    )
-    result = executeSQL(
-        sql, title="Checking if the table is a flextable.", method="fetchall",
-    )
-    return bool(result)
-
-
-def isvmap(
-    expr: Union[str, str_sql], column: str,
-):
-    """
-Checks if the input column is a VMap.
-
-Parameters
-----------
-expr: str / vDataFrame
-    Any relation or expression. If you enter an expression,
-    you must enclose it in parentheses and provide an alias.
-column: str
-    Name of the column to check.
-
-Returns
--------
-bool
-    True if the column is a VMap.
-    """
-    # -#
-    from verticapy.vdataframe import vDataFrame
-    from verticapy.connect import current_cursor
-
-    column = quote_ident(column)
-    if isinstance(expr, vDataFrame):
-        expr = expr.__genSQL__()
-    sql = f"SELECT MAPVERSION({column}) AS isvmap, {column} FROM {expr} WHERE {column} IS NOT NULL LIMIT 1;"
-    try:
-        result = executeSQL(
-            sql, title="Checking if the column is a vmap.", method="fetchall",
-        )
-        dtype = current_cursor().description[1][1]
-        if dtype not in (
-            115,
-            116,
-        ):  # 116 is for long varbinary and 115 is for long varchar
-            return False
-    except Exception as e:
-        if "'utf-8' codec can't decode byte" in str(e):
-            try:
-                sql = f"SELECT MAPVERSION({column}) AS isvmap FROM {expr} WHERE {column} IS NOT NULL LIMIT 1;"
-                result = executeSQL(
-                    sql, title="Checking if the column is a vmap.", method="fetchall",
-                )
-            except:
-                return False
-        else:
-            return False
-    if len(result) == 0 or (result[0][0] == -1):
-        return False
-    else:
-        return True
-
-
 @save_verticapy_logs
 def pandas_to_vertica(
     df: pd.DataFrame,
@@ -1054,638 +887,7 @@ read_json : Ingests a JSON file into the Vertica database.
     return vdf
 
 
-def pcsv(
-    path: str,
-    sep: str = ",",
-    header: bool = True,
-    header_names: list = [],
-    na_rep: str = "",
-    quotechar: str = '"',
-    escape: str = "\027",
-    record_terminator: str = "\n",
-    trim: bool = True,
-    omit_empty_keys: bool = False,
-    reject_on_duplicate: bool = False,
-    reject_on_empty_key: bool = False,
-    reject_on_materialized_type_error: bool = False,
-    ingest_local: bool = True,
-    flex_name: str = "",
-    genSQL: bool = False,
-):
-    """
-Parses a CSV file using flex tables. It will identify the columns and their
-respective types.
-
-Parameters
-----------
-path: str
-    Absolute path where the CSV file is located.
-sep: str, optional
-    Column separator.
-header: bool, optional
-    If set to False, the parameter 'header_names' will be to use to name the 
-    different columns.
-header_names: list, optional
-    List of the columns names.
-na_rep: str, optional
-    Missing values representation.
-quotechar: str, optional
-    Char which is enclosing the str values.
-escape: str, optional
-    Separator between each record.
-record_terminator: str, optional
-    A single-character value used to specify the end of a record.
-trim: bool, optional
-    Boolean, specifies whether to trim white space from header names and 
-    key values.
-omit_empty_keys: bool, optional
-    Boolean, specifies how the parser handles header keys without values. 
-    If true, keys with an empty value in the header row are not loaded.
-reject_on_duplicate: bool, optional
-    Boolean, specifies whether to ignore duplicate records (False), or to 
-    reject duplicates (True). In either case, the load continues.
-reject_on_empty_key: bool, optional
-    Boolean, specifies whether to reject any row containing a key without a 
-    value.
-reject_on_materialized_type_error: bool, optional
-    Boolean, specifies whether to reject any materialized column value that the 
-    parser cannot coerce into a compatible data type.
-ingest_local: bool, optional
-    If set to True, the file will be ingested from the local machine.
-flex_name: str, optional
-    Flex table name.
-genSQL: bool, optional
-    If set to True, the SQL code for creating the final table is 
-    generated but not executed. This is a good way to change the
-    final relation types or to customize the data ingestion.
-
-Returns
--------
-dict
-    dictionary containing each column and its type.
-
-See Also
---------
-read_csv  : Ingests a CSV file into the Vertica database.
-read_json : Ingests a JSON file into the Vertica database.
-    """
-    if record_terminator == "\n":
-        record_terminator = "\\n"
-    if not (flex_name):
-        flex_name = gen_tmp_name(name="flex")[1:-1]
-    if header_names:
-        header_names = f"header_names = '{sep.join(header_names)}',"
-    else:
-        header_names = ""
-    ingest_local = " LOCAL" if ingest_local else ""
-    trim = str(trim).lower()
-    omit_empty_keys = str(omit_empty_keys).lower()
-    reject_on_duplicate = str(reject_on_duplicate).lower()
-    reject_on_empty_key = str(reject_on_empty_key).lower()
-    reject_on_materialized_type_error = str(reject_on_materialized_type_error).lower()
-    compression = extract_compression(path)
-    query = f"CREATE FLEX LOCAL TEMP TABLE {flex_name}(x int) ON COMMIT PRESERVE ROWS;"
-    query2 = f"""
-       COPY {flex_name} 
-       FROM{ingest_local} '{path}' {compression} 
-       PARSER FCSVPARSER(
-            type = 'traditional', 
-            delimiter = '{sep}', 
-            header = {header}, {header_names} 
-            enclosed_by = '{quotechar}', 
-            escape = '{escape}',
-            record_terminator = '{record_terminator}',
-            trim = {trim},
-            omit_empty_keys = {omit_empty_keys},
-            reject_on_duplicate = {reject_on_duplicate},
-            reject_on_empty_key = {reject_on_empty_key},
-            reject_on_materialized_type_error = {reject_on_materialized_type_error}) 
-       NULL '{na_rep}';"""
-    if genSQL:
-        return [clean_query(query), clean_query(query2)]
-    executeSQL(
-        query=query, title="Creating flex table to identify the data types.",
-    )
-    executeSQL(
-        query=query2, title="Parsing the data.",
-    )
-    result = compute_flextable_keys(flex_name)
-    dtype = {}
-    for column_dtype in result:
-        try:
-            executeSQL(
-                query=f"""
-                    SELECT /*+LABEL('utilities.pcsv')*/
-                        (CASE 
-                            WHEN "{column_dtype[0]}"=\'{na_rep}\' THEN NULL 
-                            ELSE "{column_dtype[0]}" 
-                         END)::{column_dtype[1]} AS "{column_dtype[0]}" 
-                    FROM {flex_name} 
-                    WHERE "{column_dtype[0]}" IS NOT NULL 
-                    LIMIT 1000""",
-                print_time_sql=False,
-            )
-            dtype[column_dtype[0]] = column_dtype[1]
-        except:
-            dtype[column_dtype[0]] = "Varchar(100)"
-    drop(flex_name, method="table")
-    return dtype
-
-
 vHelp = help_start
-
-
-def pjson(path: str, ingest_local: bool = True):
-    """
-Parses a JSON file using flex tables. It will identify the columns and their
-respective types.
-
-Parameters
-----------
-path: str
-    Absolute path where the JSON file is located.
-ingest_local: bool, optional
-    If set to True, the file will be ingested from the local machine.
-
-Returns
--------
-dict
-    dictionary containing for each column its type.
-
-See Also
---------
-read_csv  : Ingests a CSV file into the Vertica database.
-read_json : Ingests a JSON file into the Vertica database.
-    """
-    flex_name = gen_tmp_name(name="flex")[1:-1]
-    executeSQL(
-        query=f"""
-            CREATE FLEX LOCAL TEMP TABLE {flex_name}
-            (x int) ON COMMIT PRESERVE ROWS;""",
-        title="Creating a flex table.",
-    )
-    path_str = path.replace("'", "''")
-    local = " LOCAL" if ingest_local else ""
-    executeSQL(
-        query=f"""
-            COPY {flex_name} FROM{local} '{path_str}' 
-            PARSER FJSONPARSER();""",
-        title="Ingesting the data.",
-    )
-    result = compute_flextable_keys(flex_name)
-    dtype = {}
-    for column_dtype in result:
-        dtype[column_dtype[0]] = column_dtype[1]
-    drop(name=flex_name, method="table")
-    return dtype
-
-
-@save_verticapy_logs
-def read_avro(
-    path: str,
-    schema: str = "",
-    table_name: str = "",
-    usecols: list = [],
-    new_name: dict = {},
-    insert: bool = False,
-    reject_on_materialized_type_error: bool = False,
-    flatten_maps: bool = True,
-    flatten_arrays: bool = False,
-    temporary_table: bool = False,
-    temporary_local_table: bool = True,
-    gen_tmp_table_name: bool = True,
-    ingest_local: bool = True,
-    genSQL: bool = False,
-    materialize: bool = True,
-    use_complex_dt: bool = False,
-):
-    """
-Ingests an AVRO file using flex tables.
-
-Parameters
-----------
-path: str
-    Absolute path where the AVRO file is located.
-schema: str, optional
-    Schema where the AVRO file will be ingested.
-table_name: str, optional
-    Final relation name.
-usecols: list, optional
-    List of the AVRO parameters to ingest. The other ones will be 
-    ignored. If empty all the AVRO parameters will be ingested.
-new_name: dict, optional
-    Dictionary of the new columns name. If the AVRO file is nested, 
-    it is advised to change the final names as special characters 
-    will be included.
-    For example, {"param": {"age": 3, "name": Badr}, "date": 1993-03-11} 
-    will create 3 columns: "param.age", "param.name" and "date". 
-    You can rename these columns using the 'new_name' parameter with 
-    the following dictionary:
-    {"param.age": "age", "param.name": "name"}
-insert: bool, optional
-    If set to True, the data will be ingested to the input relation.
-    The AVRO parameters must be the same as the input relation otherwise
-    they will not be ingested. Also, table_name cannot be empty if this is true.
-reject_on_materialized_type_error: bool, optional
-    Boolean, whether to reject a data row that contains a materialized 
-    column value that cannot be coerced into a compatible data type. 
-    If the value is false and the type cannot be coerced, the parser 
-    sets the value in that column to null.
-    If the column is a strongly-typed complex type, as opposed to a 
-    flexible complex type, then a type mismatch anywhere in the complex 
-    type causes the entire column to be treated as a mismatch. The parser 
-    does not partially load complex types.
-flatten_maps: bool, optional
-    Boolean, whether to flatten all AVRO maps. Key names are 
-    concatenated with nested levels. This value is recursive and 
-    affects all data in the load.
-flatten_arrays: bool, optional
-    Boolean, whether to convert lists to sub-maps with integer keys. 
-    When lists are flattened, key names are concatenated as for maps. 
-    Lists are not flattened by default. This value affects all data in 
-    the load, including nested lists.
-temporary_table: bool, optional
-    If set to True, a temporary table will be created.
-temporary_local_table: bool, optional
-    If set to True, a temporary local table will be created. The parameter 
-    'schema' must be empty, otherwise this parameter is ignored.
-gen_tmp_table_name: bool, optional
-    Sets the name of the temporary table. This parameter is only used when 
-    the parameter 'temporary_local_table' is set to True and if the parameters 
-    "table_name" and "schema" are unspecified.
-ingest_local: bool, optional
-    If set to True, the file will be ingested from the local machine.
-genSQL: bool, optional
-    If set to True, the SQL code for creating the final table is 
-    generated but not executed. This is a good way to change the final
-    relation types or to customize the data ingestion.
-materialize: bool, optional
-    If set to True, the flex table is materialized into a table.
-    Otherwise, it will remain a flex table. Flex tables simplify the
-    data ingestion but have worse performace compared to regular tables.
-use_complex_dt: bool, optional
-    Boolean, whether the input data file has complex structure.
-    When this is true, most of the other parameters will be ignored.
-
-Returns
--------
-vDataFrame
-    The vDataFrame of the relation.
-
-See Also
---------
-read_json : Ingests a JSON file into the Vertica database.
-    """
-    return read_json(
-        path=path,
-        schema=schema,
-        table_name=table_name,
-        usecols=usecols,
-        new_name=new_name,
-        insert=insert,
-        reject_on_materialized_type_error=reject_on_materialized_type_error,
-        flatten_maps=flatten_maps,
-        flatten_arrays=flatten_arrays,
-        temporary_table=temporary_table,
-        temporary_local_table=temporary_local_table,
-        gen_tmp_table_name=gen_tmp_table_name,
-        ingest_local=ingest_local,
-        genSQL=genSQL,
-        materialize=materialize,
-        use_complex_dt=use_complex_dt,
-        is_avro=True,
-    )
-
-
-@save_verticapy_logs
-def read_csv(
-    path: str,
-    schema: str = "",
-    table_name: str = "",
-    sep: str = "",
-    header: bool = True,
-    header_names: list = [],
-    dtype: dict = {},
-    na_rep: str = "",
-    quotechar: str = '"',
-    escape: str = "\027",
-    record_terminator: str = "\n",
-    trim: bool = True,
-    omit_empty_keys: bool = False,
-    reject_on_duplicate: bool = False,
-    reject_on_empty_key: bool = False,
-    reject_on_materialized_type_error: bool = False,
-    parse_nrows: int = -1,
-    insert: bool = False,
-    temporary_table: bool = False,
-    temporary_local_table: bool = True,
-    gen_tmp_table_name: bool = True,
-    ingest_local: bool = True,
-    genSQL: bool = False,
-    materialize: bool = True,
-):
-    """
-Ingests a CSV file using flex tables.
-
-Parameters
-----------
-path: str
-	Absolute path where the CSV file is located.
-schema: str, optional
-	Schema where the CSV file will be ingested.
-table_name: str, optional
-	The final relation/table name. If unspecified, the the name is set to the 
-    name of the file or parent directory.
-sep: str, optional
-	Column separator. 
-    If empty, the separator is guessed. This is only possible if the files
-    are not compressed.
-header: bool, optional
-	If set to False, the parameter 'header_names' will be to use to name the 
-	different columns.
-header_names: list, optional
-	List of the columns names.
-dtype: dict, optional
-    Dictionary of the user types. Providing a dictionary can increase 
-    ingestion speed and precision; instead of parsing the file to guess 
-    the different types, VerticaPy will use the input types.
-na_rep: str, optional
-	Missing values representation.
-quotechar: str, optional
-	Char which is enclosing the str values.
-escape: str, optional
-	Separator between each record.
-record_terminator: str, optional
-    A single-character value used to specify the end of a record.
-trim: bool, optional
-    Boolean, specifies whether to trim white space from header names and 
-    key values.
-omit_empty_keys: bool, optional
-    Boolean, specifies how the parser handles header keys without values. 
-    If true, keys with an empty value in the header row are not loaded.
-reject_on_duplicate: bool, optional
-    Boolean, specifies whether to ignore duplicate records (False), or to 
-    reject duplicates (True). In either case, the load continues.
-reject_on_empty_key: bool, optional
-    Boolean, specifies whether to reject any row containing a key without a 
-    value.
-reject_on_materialized_type_error: bool, optional
-    Boolean, specifies whether to reject any materialized column value that the 
-    parser cannot coerce into a compatible data type.
-parse_nrows: int, optional
-	If this parameter is greater than 0. A new file of 'parse_nrows' rows
-	will be created and ingested first to identify the data types. It will be
-	then dropped and the entire file will be ingested. The data types identification
-	will be less precise but this parameter can make the process faster if the
-	file is heavy.
-insert: bool, optional
-	If set to True, the data will be ingested to the input relation. Be sure
-	that your file has a header corresponding to the name of the relation
-	columns, otherwise ingestion will fail.
-temporary_table: bool, optional
-    If set to True, a temporary table will be created.
-temporary_local_table: bool, optional
-    If set to True, a temporary local table will be created. The parameter 'schema'
-    must be empty, otherwise this parameter is ignored.
-gen_tmp_table_name: bool, optional
-    Sets the name of the temporary table. This parameter is only used when the 
-    parameter 'temporary_local_table' is set to True and if the parameters 
-    "table_name" and "schema" are unspecified.
-ingest_local: bool, optional
-    If set to True, the file will be ingested from the local machine.
-genSQL: bool, optional
-    If set to True, the SQL code for creating the final table is 
-    generated but not executed. This is a good way to change the final
-    relation types or to customize the data ingestion.
-materialize: bool, optional
-    If set to True, the flex table is materialized into a table.
-    Otherwise, it will remain a flex table. Flex tables simplify the
-    data ingestion but have worse performace compared to regular tables.
-
-Returns
--------
-vDataFrame
-	The vDataFrame of the relation.
-
-See Also
---------
-read_json : Ingests a JSON file into the Vertica database.
-	"""
-    from verticapy import vDataFrame
-
-    if schema:
-        temporary_local_table = False
-    elif temporary_local_table:
-        schema = "v_temp_schema"
-    else:
-        schema = "public"
-    if header_names and dtype:
-        warning_message = (
-            "Parameters 'header_names' and 'dtype' are both defined. "
-            "Only 'dtype' will be used."
-        )
-        warnings.warn(warning_message, Warning)
-    basename = ".".join(path.split("/")[-1].split(".")[0:-1])
-    if gen_tmp_table_name and temporary_local_table and not (table_name):
-        table_name = gen_tmp_name(name=basename)
-    assert not (temporary_table) or not (temporary_local_table), ParameterError(
-        "Parameters 'temporary_table' and 'temporary_local_table' can not be both "
-        "set to True."
-    )
-    path, sep, header_names, na_rep, quotechar, escape = (
-        path.replace("'", "''"),
-        sep.replace("'", "''"),
-        [str(elem).replace("'", "''") for elem in header_names],
-        na_rep.replace("'", "''"),
-        quotechar.replace("'", "''"),
-        escape.replace("'", "''"),
-    )
-    file_extension = path.split(".")[-1].lower()
-    compression = extract_compression(path)
-    if file_extension != "csv" and (compression == "UNCOMPRESSED"):
-        raise ExtensionError("The file extension is incorrect !")
-    multiple_files = False
-    if "*" in basename:
-        multiple_files = True
-    if not (genSQL):
-        table_name_str = table_name.replace("'", "''")
-        schema_str = schema.replace("'", "''")
-        result = executeSQL(
-            query=f"""
-                SELECT /*+LABEL('utilities.read_csv')*/
-                    column_name 
-               FROM columns 
-               WHERE table_name = '{table_name_str}' 
-                 AND table_schema = '{schema_str}' 
-               ORDER BY ordinal_position""",
-            title="Looking if the relation exists.",
-            method="fetchall",
-        )
-    input_relation = format_schema_table(schema, table_name)
-    if not (genSQL) and (result != []) and not (insert) and not (genSQL):
-        raise NameError(f"The table {input_relation} already exists !")
-    elif not (genSQL) and (result == []) and (insert):
-        raise MissingRelation(f"The table {input_relation} doesn't exist !")
-    else:
-        if temporary_local_table:
-            input_relation = f"v_temp_schema.{quote_ident(table_name)}"
-        file_header = []
-        path_first_file_in_folder = path
-        if multiple_files and ingest_local:
-            path_first_file_in_folder = get_first_file(path, "csv")
-        if (
-            not (header_names)
-            and not (dtype)
-            and (compression == "UNCOMPRESSED")
-            and ingest_local
-        ):
-            if not (path_first_file_in_folder):
-                raise ParameterError("No CSV file detected in the folder.")
-            file_header = get_header_name_csv(path_first_file_in_folder, sep)
-        elif not (header_names) and not (dtype) and (compression != "UNCOMPRESSED"):
-            raise ParameterError(
-                "The input file is compressed and parameters 'dtypes' and 'header_names'"
-                " are not defined. It is impossible to read the file's header."
-            )
-        elif not (header_names) and not (dtype) and not (ingest_local):
-            raise ParameterError(
-                "The input file is in the Vertica server and parameters 'dtypes' and "
-                "'header_names' are not defined. It is impossible to read the file's header."
-            )
-        if (header_names == []) and (header):
-            if not (dtype):
-                header_names = file_header
-            else:
-                header_names = [elem for elem in dtype]
-            header_names = erase_space_start_end_in_list_values(header_names)
-        elif len(file_header) > len(header_names):
-            header_names += [
-                f"ucol{i + len(header_names)}"
-                for i in range(len(file_header) - len(header_names))
-            ]
-        if not (sep):
-            try:
-                f = open(path_first_file_in_folder, "r")
-                file_str = f.readline()
-                f.close()
-                sep = guess_sep(file_str)
-            except:
-                sep = ","
-        if not (materialize):
-            suffix, prefix, final_relation = (
-                "",
-                " ON COMMIT PRESERVE ROWS;",
-                input_relation,
-            )
-            if temporary_local_table:
-                suffix = "LOCAL TEMP "
-                final_relation = table_name
-            elif temporary_table:
-                suffix = "TEMP "
-            else:
-                prefix = ";"
-            query = f"CREATE FLEX {suffix}TABLE {final_relation}(x int){prefix}"
-            query2 = pcsv(
-                path=path,
-                sep=sep,
-                header=header,
-                header_names=header_names,
-                na_rep=na_rep,
-                quotechar=quotechar,
-                escape=escape,
-                record_terminator=record_terminator,
-                trim=trim,
-                omit_empty_keys=omit_empty_keys,
-                reject_on_duplicate=reject_on_duplicate,
-                reject_on_empty_key=reject_on_empty_key,
-                reject_on_materialized_type_error=reject_on_materialized_type_error,
-                ingest_local=ingest_local,
-                flex_name=input_relation,
-                genSQL=True,
-            )[1]
-            if genSQL and not (insert):
-                return [clean_query(query), clean_query(query2)]
-            elif genSQL:
-                return [clean_query(query2)]
-            if not (insert):
-                executeSQL(
-                    query, title="Creating the flex table.",
-                )
-            executeSQL(
-                query2, title="Copying the data.",
-            )
-            return vDataFrame(table_name, schema=schema)
-        if (
-            (parse_nrows > 0)
-            and not (insert)
-            and (compression == "UNCOMPRESSED")
-            and ingest_local
-        ):
-            f = open(path_first_file_in_folder, "r")
-            path_test = path_first_file_in_folder.split(".")[-2] + "_verticapy_copy.csv"
-            f2 = open(path_test, "w")
-            for i in range(parse_nrows + int(header)):
-                line = f.readline()
-                f2.write(line)
-            f.close()
-            f2.close()
-        else:
-            path_test = path_first_file_in_folder
-        query1 = ""
-        if not (insert):
-            if not (dtype):
-                dtype = pcsv(
-                    path_test,
-                    sep,
-                    header,
-                    header_names,
-                    na_rep,
-                    quotechar,
-                    escape,
-                    ingest_local=ingest_local,
-                )
-            if parse_nrows > 0:
-                os.remove(path_test)
-            dtype_sorted = {}
-            for elem in header_names:
-                key = find_val_in_dict(elem, dtype, return_key=True)
-                dtype_sorted[key] = dtype[key]
-            query1 = create_table(
-                table_name,
-                dtype_sorted,
-                schema,
-                temporary_table,
-                temporary_local_table,
-                genSQL=True,
-            )
-        skip = " SKIP 1" if (header) else ""
-        local = "LOCAL " if ingest_local else ""
-        header_names_str = ", ".join([f'"{column}"' for column in header_names])
-        query2 = f"""
-            COPY {input_relation}({header_names_str}) 
-            FROM {local}'{path}' {compression} 
-            DELIMITER '{sep}' 
-            NULL '{na_rep}' 
-            ENCLOSED BY '{quotechar}' 
-            ESCAPE AS '{escape}'{skip};"""
-        if genSQL:
-            if insert:
-                return [clean_query(query2)]
-            else:
-                return [clean_query(query1), clean_query(query2)]
-        else:
-            if not (insert):
-                executeSQL(query1, title="Creating the table.")
-            executeSQL(
-                query2, title="Ingesting the data.",
-            )
-            if (
-                not (insert)
-                and not (temporary_local_table)
-                and vp.OPTIONS["print_info"]
-            ):
-                print(f"The table {input_relation} has been successfully created.")
-            return vDataFrame(table_name, schema=schema)
 
 
 @check_minimum_version
@@ -1897,376 +1099,6 @@ vDataFrame
 
 
 @save_verticapy_logs
-def read_json(
-    path: str,
-    schema: str = "",
-    table_name: str = "",
-    usecols: list = [],
-    new_name: dict = {},
-    insert: bool = False,
-    start_point: str = None,
-    record_terminator: str = None,
-    suppress_nonalphanumeric_key_chars: bool = False,
-    reject_on_materialized_type_error: bool = False,
-    reject_on_duplicate: bool = False,
-    reject_on_empty_key: bool = False,
-    flatten_maps: bool = True,
-    flatten_arrays: bool = False,
-    temporary_table: bool = False,
-    temporary_local_table: bool = True,
-    gen_tmp_table_name: bool = True,
-    ingest_local: bool = True,
-    genSQL: bool = False,
-    materialize: bool = True,
-    use_complex_dt: bool = False,
-    is_avro: bool = False,
-):
-    """
-Ingests a JSON file using flex tables.
-
-Parameters
-----------
-path: str
-	Absolute path where the JSON file is located.
-schema: str, optional
-	Schema where the JSON file will be ingested.
-table_name: str, optional
-	Final relation name.
-usecols: list, optional
-	List of the JSON parameters to ingest. The other ones will be 
-    ignored. If empty all the JSON parameters will be ingested.
-new_name: dict, optional
-	Dictionary of the new columns name. If the JSON file is nested, 
-    it is advised to change the final names as special characters 
-    will be included.
-	For example, {"param": {"age": 3, "name": Badr}, "date": 1993-03-11} 
-    will create 3 columns: "param.age", "param.name" and "date". 
-    You can rename these columns using the 'new_name' parameter with 
-    the following dictionary:
-	{"param.age": "age", "param.name": "name"}
-insert: bool, optional
-	If set to True, the data will be ingested to the input relation.
-    The JSON parameters must be the same as the input relation otherwise
-    they will not be ingested. Also, table_name cannot be empty if this is true.
-start_point: str, optional
-    String, name of a key in the JSON load data at which to begin parsing. 
-    The parser ignores all data before the start_point value. 
-    The value is loaded for each object in the file. The parser processes 
-    data after the first instance, and up to the second, ignoring any 
-    remaining data.
-record_terminator: str, optional
-    When set, any invalid JSON records are skipped and parsing continues 
-    with the next record. 
-    Records must be terminated uniformly. For example, if your input file 
-    has JSON records terminated by newline characters, set this parameter 
-    to '\n'). 
-    If any invalid JSON records exist, parsing continues after the next 
-    record_terminator.
-    Even if the data does not contain invalid records, specifying an 
-    explicit record terminator can improve load performance by allowing 
-    cooperative parse and apportioned load to operate more efficiently.
-    When you omit this parameter, parsing ends at the first invalid JSON 
-    record.
-suppress_nonalphanumeric_key_chars: bool, optional
-    Boolean, whether to suppress non-alphanumeric characters in JSON 
-    key values. The parser replaces these characters with an underscore 
-    (_) when this parameter is true.
-reject_on_materialized_type_error: bool, optional
-    Boolean, whether to reject a data row that contains a materialized 
-    column value that cannot be coerced into a compatible data type. 
-    If the value is false and the type cannot be coerced, the parser 
-    sets the value in that column to null.
-    If the column is a strongly-typed complex type, as opposed to a 
-    flexible complex type, then a type mismatch anywhere in the complex 
-    type causes the entire column to be treated as a mismatch. The parser 
-    does not partially load complex types.
-reject_on_duplicate: bool, optional
-    Boolean, whether to ignore duplicate records (false), or to 
-    reject duplicates (true). In either case, the load continues.
-reject_on_empty_key: bool, optional
-    Boolean, whether to reject any row containing a field key 
-    without a value.
-flatten_maps: bool, optional
-    Boolean, whether to flatten sub-maps within the JSON data, separating map levels 
-    with a period (.). This value affects all data in the load, including nested maps.
-flatten_arrays: bool, optional
-    Boolean, whether to convert lists to sub-maps with integer keys. 
-    When lists are flattened, key names are concatenated as for maps. 
-    Lists are not flattened by default. This value affects all data in 
-    the load, including nested lists.
-temporary_table: bool, optional
-    If set to True, a temporary table will be created.
-temporary_local_table: bool, optional
-    If set to True, a temporary local table will be created. The parameter 
-    'schema' must be empty, otherwise this parameter is ignored.
-gen_tmp_table_name: bool, optional
-    Sets the name of the temporary table. This parameter is only used when 
-    the parameter 'temporary_local_table' is set to True and if the parameters 
-    "table_name" and "schema" are unspecified.
-ingest_local: bool, optional
-    If set to True, the file will be ingested from the local machine.
-genSQL: bool, optional
-    If set to True, the SQL code for creating the final table is 
-    generated but not executed. This is a good way to change the final
-    relation types or to customize the data ingestion.
-materialize: bool, optional
-    If set to True, the flex table is materialized into a table.
-    Otherwise, it will remain a flex table. Flex tables simplify the
-    data ingestion but have worse performace compared to regular tables.
-use_complex_dt: bool, optional
-    Boolean, whether the input data file has complex structure.
-    When this is true, most of the other parameters will be ignored.
-
-Returns
--------
-vDataFrame
-	The vDataFrame of the relation.
-
-See Also
---------
-read_csv : Ingests a CSV file into the Vertica database.
-	"""
-    from verticapy import vDataFrame
-
-    if use_complex_dt:
-        assert not (new_name), ParameterError(
-            "You cannot use the parameter " "new_name" " with " "use_complex_dt" "."
-        )
-        if is_avro:
-            max_files = 1
-        elif ("*" in path) and ingest_local:
-            dirname = os.path.dirname(path)
-            all_files = os.listdir(dirname)
-            max_files = sum(1 for x in all_files if x.endswith(".json"))
-        else:
-            max_files = 1000
-        return read_file(
-            path=path,
-            schema=schema,
-            table_name=table_name,
-            insert=insert,
-            temporary_table=temporary_table,
-            temporary_local_table=temporary_local_table,
-            gen_tmp_table_name=gen_tmp_table_name,
-            ingest_local=ingest_local,
-            genSQL=genSQL,
-            max_files=max_files,
-        )
-    if schema:
-        temporary_local_table = False
-    elif temporary_local_table:
-        schema = "v_temp_schema"
-    else:
-        schema = "public"
-    assert not (temporary_table) or not (temporary_local_table), ParameterError(
-        "Parameters 'temporary_table' and 'temporary_local_table' can not be both set to True."
-    )
-    file_extension = path.split(".")[-1].lower()
-    compression = extract_compression(path)
-    if (
-        (file_extension not in ("json",) and not (is_avro))
-        or (file_extension not in ("avro",) and (is_avro))
-    ) and (compression == "UNCOMPRESSED"):
-        raise ExtensionError("The file extension is incorrect !")
-    basename = ".".join(path.split("/")[-1].split(".")[0:-1])
-    if gen_tmp_table_name and temporary_local_table and not (table_name):
-        table_name = gen_tmp_name(name=basename)
-    if not (table_name):
-        table_name = basename
-    if is_avro:
-        label = "read_avro"
-        parser = "FAVROPARSER"
-    else:
-        label = "read_json"
-        parser = "FJSONPARSER"
-    if not (genSQL):
-        table_name_str = table_name.replace("'", "''")
-        schema_str = schema.replace("'", "''")
-        column_name = executeSQL(
-            query=f"""
-                SELECT 
-                    /*+LABEL('utilities.{label}')*/ 
-                    column_name,
-                    data_type 
-                FROM columns 
-                WHERE table_name = '{table_name_str}' 
-                  AND table_schema = '{schema_str}' 
-                ORDER BY ordinal_position""",
-            title="Looking if the relation exists.",
-            method="fetchall",
-        )
-    input_relation = format_schema_table(schema, table_name)
-    if not (genSQL) and (column_name != []) and not (insert):
-        raise NameError(f"The table {input_relation} already exists !")
-    elif not (genSQL) and (column_name == []) and (insert):
-        raise MissingRelation(f"The table {input_relation} doesn't exist !")
-    else:
-        if temporary_local_table:
-            input_relation = quote_ident(table_name)
-        all_queries = []
-        if not (materialize):
-            suffix, prefix = "", "ON COMMIT PRESERVE ROWS;"
-            if temporary_local_table:
-                suffix = "LOCAL TEMP "
-            elif temporary_table:
-                suffix = "TEMP "
-            else:
-                prefix = ";"
-            query = f"""
-                CREATE FLEX {suffix}TABLE 
-                {input_relation}(x int){prefix}"""
-        else:
-            flex_name = gen_tmp_name(name="flex")[1:-1]
-            query = f"""
-                CREATE FLEX LOCAL TEMP TABLE {flex_name}(x int) 
-                ON COMMIT PRESERVE ROWS;"""
-        if not (insert):
-            all_queries += [clean_query(query)]
-        options = []
-        if start_point and not (is_avro):
-            options += [f"start_point='{start_point}'"]
-        if record_terminator and not (is_avro):
-            prefix = ""
-            if "\\" in record_terminator.__repr__():
-                prefix = "E"
-            options += [f"record_terminator={prefix}'{record_terminator}'"]
-        if suppress_nonalphanumeric_key_chars and not (is_avro):
-            options += ["suppress_nonalphanumeric_key_chars=true"]
-        else:
-            options += ["suppress_nonalphanumeric_key_chars=false"]
-        if reject_on_materialized_type_error:
-            assert materialize, ParameterError(
-                "When using complex data types the table has to "
-                "be materialized. Set materialize to True"
-            )
-            options += ["reject_on_materialized_type_error=true"]
-        else:
-            options += ["reject_on_materialized_type_error=false"]
-        if reject_on_duplicate and not (is_avro):
-            options += ["reject_on_duplicate=true"]
-        else:
-            options += ["reject_on_duplicate=false"]
-        if reject_on_empty_key and not (is_avro):
-            options += ["reject_on_empty_key=true"]
-        else:
-            options += ["reject_on_empty_key=false"]
-        if flatten_arrays:
-            options += ["flatten_arrays=true"]
-        else:
-            options += ["flatten_arrays=false"]
-        if flatten_maps:
-            options += ["flatten_maps=true"]
-        else:
-            options += ["flatten_maps=false"]
-        materialize_str = flex_name if (materialize) else input_relation
-        local = " LOCAL" if ingest_local else ""
-        path_str = path.replace("'", "''")
-        query2 = f"""
-            COPY {materialize_str} 
-            FROM{local} '{path_str}' {compression} 
-            PARSER {parser}({", ".join(options)});"""
-        all_queries = all_queries + [clean_query(query2)]
-        if genSQL and insert and not (materialize):
-            return [clean_query(query2)]
-        elif genSQL and not (materialize):
-            return all_queries
-        if not (insert):
-            executeSQL(
-                query, title="Creating flex table.",
-            )
-        executeSQL(
-            query2, title="Ingesting the data in the flex table.",
-        )
-        if not (materialize):
-            return vDataFrame(table_name, schema=schema)
-        result = compute_flextable_keys(flex_name)
-        dtype = {}
-        for column_dtype in result:
-            try:
-                executeSQL(
-                    query=f"""
-                        SELECT 
-                            /*+LABEL('utilities.{label}')*/ 
-                            \"{column_dtype[0]}\"::{column_dtype[1]} 
-                        FROM {flex_name} 
-                        LIMIT 1000""",
-                    print_time_sql=False,
-                )
-                dtype[column_dtype[0]] = column_dtype[1]
-            except:
-                dtype[column_dtype[0]] = "Varchar(100)"
-        if not (insert):
-            cols = (
-                [column for column in dtype]
-                if not (usecols)
-                else [column for column in usecols]
-            )
-            for i, column in enumerate(cols):
-                column_str = column.replace('"', "")
-                if column in new_name:
-                    cols[i] = f'"{column_str}"::{dtype[column]} AS "{new_name[column]}"'
-                else:
-                    cols[i] = f'"{column_str}"::{dtype[column]}'
-            if temporary_local_table:
-                suffix = "LOCAL TEMPORARY "
-            elif temporary_table:
-                suffix = "TEMPORARY "
-            else:
-                suffix = ""
-            on_commit = " ON COMMIT PRESERVE ROWS" if suffix else ""
-            query3 = f"""
-                CREATE {suffix}TABLE {input_relation}{on_commit} AS 
-                    SELECT 
-                        /*+LABEL('utilities.{label}')*/ 
-                        {", ".join(cols)} 
-                    FROM {flex_name}"""
-            all_queries = all_queries + [clean_query(query3)]
-            if genSQL:
-                return all_queries
-            executeSQL(
-                query3, title="Creating table.",
-            )
-            if not (temporary_local_table) and vp.OPTIONS["print_info"]:
-                print(f"The table {input_relation} has been successfully created.")
-        else:
-            column_name_dtype = {}
-            for elem in column_name:
-                column_name_dtype[elem[0]] = elem[1]
-            final_cols = {}
-            for column in column_name_dtype:
-                final_cols[column] = None
-            for column in column_name_dtype:
-                if column in dtype:
-                    final_cols[column] = column
-                else:
-                    for col in new_name:
-                        if new_name[col] == column:
-                            final_cols[column] = col
-            final_transformation = []
-            for column in final_cols:
-                if final_cols[column] == None:
-                    final_transformation += [f'NULL AS "{column}"']
-                else:
-                    final_transformation += [
-                        f'"{final_cols}"::{column_name_dtype[column]} AS "{column}"'
-                    ]
-            query = f"""
-                INSERT 
-                    /*+LABEL('utilities.{label}')*/ 
-                INTO {input_relation} 
-                SELECT 
-                    {", ".join(final_transformation)} 
-                FROM {flex_name}"""
-            if genSQL:
-                return [clean_query(query)]
-            executeSQL(
-                query, title="Inserting data into table.",
-            )
-        drop(name=flex_name, method="table")
-        return vDataFrame(table_name, schema=schema)
-
-
-@save_verticapy_logs
 def read_shp(
     path: str, schema: str = "public", table_name: str = "",
 ):
@@ -2464,193 +1296,6 @@ bool
         return True
     except:
         return False
-
-
-@overload
-def set_option(
-    option: Literal["color_style"], value: Literal[tuple(vp_colors.COLORS_OPTIONS)]
-) -> None:
-    ...
-
-
-@overload
-def set_option(option: Literal["mode"], value: Literal["light", "full"]) -> None:
-    ...
-
-
-def set_option(
-    option: Literal[
-        "cache",
-        "colors",
-        "color_style",
-        "interactive",
-        "count_on",
-        "footer_on",
-        "max_columns",
-        "max_rows",
-        "mode",
-        "overwrite_model",
-        "percent_bar",
-        "print_info",
-        "random_state",
-        "save_query_profile",
-        "sql_on",
-        "temp_schema",
-        "time_on",
-        "tqdm",
-    ],
-    value: Union[bool, int, str, list, None] = None,
-) -> None:
-    """
-    Sets VerticaPy options.
-
-    Parameters
-    ----------
-    option: str
-        Option to use.
-        cache              : bool
-            If set to True, the vDataFrame will save in memory the computed
-            aggregations.
-        colors             : list
-            List of the colors used to draw the graphics.
-        color_style        : str
-            Style used to color the graphics, one of the following:
-            "rgb", "sunset", "retro", "shimbg", "swamp", "med", "orchid", 
-            "magenta", "orange", "vintage", "vivid", "berries", "refreshing", 
-            "summer", "tropical", "india", "default".
-        count_on           : bool
-            If set to True, the total number of rows in vDataFrames and tablesamples is  
-            computed and displayed in the footer (if footer_on is True).
-        footer_on          : bool
-            If set to True, vDataFrames and tablesamples show a footer that includes information 
-            about the displayed rows and columns.
-        interactive        : bool
-            If set to True, verticaPy outputs will be displayed on interactive tables. 
-        max_columns        : int
-            Maximum number of columns to display. If the parameter is incorrect, 
-            nothing is changed.
-        max_rows           : int
-            Maximum number of rows to display. If the parameter is incorrect, 
-            nothing is changed.
-        mode               : str
-            How to display VerticaPy outputs.
-                full  : VerticaPy regular display mode.
-                light : Minimalist display mode.
-        overwrite_model: bool
-            If set to True and you try to train a model with an existing name. 
-            It will be automatically overwritten.
-        percent_bar        : bool
-            If set to True, it displays the percent of non-missing values.
-        print_info         : bool
-            If set to True, information will be printed each time the vDataFrame 
-            is modified.
-        random_state       : int
-            Integer used to seed the random number generation in VerticaPy.
-        save_query_profile : str / list / bool
-            If set to "all" or True, all function calls are stored in the query 
-            profile table. This makes it possible to differentiate the VerticaPy 
-            logs from the Vertica logs.
-            You can also provide a list of specific methods to store. For example: 
-            if you specify ["corr", "describe"], only the logs associated with 
-            those two methods are stored. 
-            If set to False, this functionality is deactivated.
-        sql_on             : bool
-            If set to True, displays all the SQL queries.
-        temp_schema        : str
-            Specifies the temporary schema that certain methods/functions use to 
-            create intermediate objects, if needed. 
-        time_on            : bool
-            If set to True, displays all the SQL queries elapsed time.
-        tqdm               : bool
-            If set to True, a loading bar is displayed when using iterative 
-            functions.
-    value: object, optional
-        New value of option.
-    """
-    wrong_value = False
-    if option == "colors":
-        if isinstance(value, list):
-            vp.OPTIONS["colors"] = [str(elem) for elem in value]
-        else:
-            wrong_value = True
-    elif option == "color_style":
-        if value == None:
-            value = "default"
-        if value in vp_colors.COLORS_OPTIONS:
-            vp.OPTIONS["colors"] = vp_colors.COLORS_OPTIONS[value]
-        else:
-            wrong_value = True
-    elif option == "max_columns":
-        if isinstance(value, int) and value > 0:
-            vp.OPTIONS["max_columns"] = int(value)
-        else:
-            wrong_value = True
-    elif option == "max_rows":
-        if isinstance(value, int) and value >= 0:
-            vp.OPTIONS["max_rows"] = int(value)
-        else:
-            wrong_value = True
-    elif option == "mode":
-        if value in ["light", "full"]:
-            vp.OPTIONS["mode"] = value
-        else:
-            wrong_value = True
-    elif option == "random_state":
-        if isinstance(value, int) and (value < 0):
-            raise ParameterError("Random State Value must be positive.")
-        if isinstance(value, int):
-            vp.OPTIONS["random_state"] = value
-        elif value == None:
-            vp.OPTIONS["random_state"] = None
-        else:
-            wrong_value = True
-    elif option in (
-        "print_info",
-        "sql_on",
-        "time_on",
-        "count_on",
-        "cache",
-        "footer_on",
-        "tqdm",
-        "overwrite_model",
-        "percent_bar",
-        "interactive",
-    ):
-        if value in (True, False, None):
-            vp.OPTIONS[option] = value
-        else:
-            wrong_value = True
-    elif option == "save_query_profile":
-        if value == "all":
-            value = True
-        elif isinstance(value, (bool, list)):
-            pass
-        else:
-            wrong_value = True
-        if not (wrong_value):
-            vp.OPTIONS[option] = value
-    elif option == "temp_schema":
-        if isinstance(value, str):
-            value_str = value.replace("'", "''")
-            query = f"""
-                SELECT /*+LABEL('utilities.set_option')*/
-                  schema_name 
-               FROM v_catalog.schemata 
-               WHERE schema_name = '{value_str}' LIMIT 1;"""
-            res = executeSQL(
-                query, title="Checking if the schema exists.", method="fetchrow"
-            )
-            if res:
-                vp.OPTIONS["temp_schema"] = str(value)
-            else:
-                raise ParameterError(f"The schema '{value}' could not be found.")
-        else:
-            wrong_value = True
-    else:
-        raise ParameterError(f"Option '{option}' does not exist.")
-    if wrong_value:
-        warning_message = "The parameter value is incorrect. Nothing was changed."
-        warnings.warn(warning_message, Warning)
 
 
 class tablesample:
